Finished basic PyQT5 tutorial
Finished the basic pat of the PyQT5 tutorial and along with it created a
simple calculator using the basic knowhow gained. Though the calc is
extremely basic, I'm looking forward to research on how other widgets
work and more about events. For now, I will create a better calculator
and wrap one of my tools using pyqt5 for example nameep, would be
amazing if it had a gui, wouldn't it? Therefore, before getting into
chapter 13 of the PyQT5 book, I will create some basic apps for learning
purposes with which I hope to understand more widgets and the pyqt5
ecosystem. Night.